backend/requirements.txt
fastapi>=0.110.0
uvicorn==0.24.0
pydantic>=2.5.2
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4
websockets>=13.0
redis==5.0.1
celery==5.3.6
asyncpg==0.29.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pytest==7.4.3
apscheduler==3.10.4
google-genai
python-dotenv
flower
